requirements.txt
django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
tweepy==4.14.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
xgboost==2.0.3
python-dotenv==1.0.1
